fix(conf_loader) default `nginx_user` to `kong kong` instead of nobody
If we install and run Kong with everything default, whenever Kong tries to write something to the Kong prefix, things will break and we will get a permission denied error. This applies to:

* Data Plane nodes trying to write the `config.json.gz` file to the Kong prefix
* Go pluginserver trying to write the `go.lua` file write to the Kong prefix
* Lua plugins trying to write files to the Kong prefix

And other cases. This is because, by default, the Kong prefix is owned by `root`. When we do a `apt-get install` or `yum install`, the `/usr/local/kong/` prefix (default) is owned by `root` by default. After that, if we do a `kong start` without changing our operating system's `$USER`, the nginx master process will be run as `root` and the nginx workers processes will be run as `nobody`. It means that if we do everything default, the nginx workers processes can’t write anything to the Kong prefix and we’ll get the permission denied errors.

The idea of this PR is to fix the permission denied errors when running Kong with the default settings without causing breaking changes. If the `kong` user/group does not exist, the user will be dropped to `nobody` as it was before. If the `nginx_user` configuration is set to something else other than `kong` (now the default), we won't change anything and the configuration value will be preserved. If it's set to `nobody` the behavior will be the same as before.

See also:
[1] https://github.com/Kong/kong-build-tools/pull/339
[2] https://github.com/Kong/kong/pull/6422/files

Co-authored-by: Datong Sun <datong.sun@konghq.com>